Add startAndEndDate utility for reports
