+++ packages/lesmis/src/utils/index.js
@@ -20,3 +20,4 @@ export * from './useEntityBreadcrumbs';
 export * from './useDropdownOptionsWithFavouriteDashboardItems';
 export * from './getExportableSubDashboards';
 export * from './export';
+export * from './useStartAndEndDates';
